+++ src/StateSmithTest/PlantUMLTests/ParsingTests.cs
@@ -28,6 +28,39 @@ public class ParsingTests
     }
 
     [Fact]
+    public void NoDiagramName()
+    {
+        ParseAssertNoError("""
+            @startuml
+
+            @enduml
+            """);
+        translator.Root.id.Should().Be("foo");
+    }
+
+    [Fact]
+    public void InvalidDiagramName()
+    {
+        ParseAssertNoError("%$foo##$*.puml", """
+            @startuml
+
+            @enduml
+            """);
+        translator.Root.id.Should().Be("foo");
+    }
+
+    [Fact]
+    public void InvalidDiagramNameStartsWithNumber()
+    {
+        ParseAssertNoError("22# $%foo##$*.puml", """
+            @startuml
+
+            @enduml
+            """);
+        translator.Root.id.Should().Be("sm22foo");
+    }
+
+    [Fact]
     public void DiagramNameIndented()
     {
         ParseAssertNoError("""
@@ -88,9 +121,6 @@ public class ParsingTests
             @enduml
             """);
 
-        // note that column is 1 because of slight hack for https://github.com/StateSmith/StateSmith/issues/352
-        action.Should().Throw<Exception>()
-            .WithMessage("PlantUML diagrams need a name and should start like `@startuml MySmName`. Location Details { line: 1, column: 1, text: `@startuml`. }");
     }
 
     [Fact]
@@ -250,13 +280,18 @@ public class ParsingTests
 
     private void ParseAssertNoError(string input)
     {
-        translator.ParseDiagramText(input);
+        ParseAssertNoError("foo.puml", input);
+    }
+
+    private void ParseAssertNoError(string filename, string input)
+    {
+        translator.ParseDiagramText(filename, input);
         translator.HasError().Should().BeFalse();
     }
 
     private void ParseAssertHasAtLeastOneError(string input)
     {
-        translator.ParseDiagramText(input);
+        translator.ParseDiagramText("foo.puml", input);
         translator.HasError().Should().BeTrue();
     }
 
@@ -336,7 +371,7 @@ public class ParsingTests
             @enduml
             """;
         InputSmBuilder inputSmBuilder = new();
-        inputSmBuilder.ConvertPlantUmlTextNodesToVertices(plantUmlText);
+        inputSmBuilder.ConvertPlantUmlTextNodesToVertices("foo.puml", plantUmlText);
         inputSmBuilder.FinishRunning();
 
         StateMachine root = inputSmBuilder.GetStateMachine();
@@ -377,7 +412,7 @@ public class ParsingTests
             @enduml
             """;
         InputSmBuilder inputSmBuilder = new();
-        inputSmBuilder.ConvertPlantUmlTextNodesToVertices(plantUmlText);
+        inputSmBuilder.ConvertPlantUmlTextNodesToVertices("foo.puml", plantUmlText);
         inputSmBuilder.FinishRunning();
 
         StateMachine root = inputSmBuilder.GetStateMachine();
@@ -507,7 +542,7 @@ public class ParsingTests
             @enduml
             """;
         InputSmBuilder inputSmBuilder = new();
-        inputSmBuilder.ConvertPlantUmlTextNodesToVertices(plantUmlText);
+        inputSmBuilder.ConvertPlantUmlTextNodesToVertices("foo.puml", plantUmlText);
         inputSmBuilder.FinishRunning();
 
         var sm = inputSmBuilder.GetStateMachine();
@@ -895,7 +930,7 @@ public class ParsingTests
             """;
 
         InputSmBuilder inputSmBuilder = new();
-        inputSmBuilder.ConvertPlantUmlTextNodesToVertices(plantUmlText);
+        inputSmBuilder.ConvertPlantUmlTextNodesToVertices("foo.puml", plantUmlText);
         inputSmBuilder.FinishRunning();
 
         StateMachine root = inputSmBuilder.GetStateMachine();